+++ assets/js/stores/CountryStore.js
@@ -22,7 +22,7 @@ export const useCountryStore = defineStore('country', {
             if(Array.isArray(apiResult)){
                 result = apiResult;
             }else{
-                result = apiResult.data;
+                result = apiResult.member || apiResult.data;
                 if(!Array.isArray(result)){
                     result = [result];
                 }
@@ -88,7 +88,9 @@ export const useCountryStore = defineStore('country', {
             this.error = null
             try {
                 const res = await CountryService.find(url)
+                debugger;
                 this.data = this.mapCountries(res);
+                debugger;
             } catch (err) {
                 this.error = err.message || 'Error fetching countries'
             } finally {